Long parameters can only be parsed into a long variable.
Using an integer causes overflow.

+++ client.cpp
@@ -291,7 +291,7 @@ PHP_METHOD(VoltClient, connect)
     char *hostname = NULL;
     char *username = NULL;
     char *password = NULL;
-    int port = 21212;
+    long port = 21212;
 
     voltdb::errType err = voltdb::errOk;
     zval *zobj = getThis();
